feat: socket response type 구분
- statusCode 설명 by backend
    -200: 성공
    -400: 잘못된 요청 예시 -> 해당 방에 없는 사람을 exit 요청하는 경우
    -403: 권한이 없을경우 예시-> 방장이 아닐경우
    -404: 찾을 수 없는 리소스 요청 예시 -> 없는 방 ID로 요청하는 경우
    -423: 리소스가 가득 찬 경우 예시 -> 해당 방의 인원이 가득 찼는데 join 하는 경우, 자금이 부족한데 핸드캡을 해당 금액 이상 주려고 하는 경우
    -500: internal server error -> 서버에 정확한 에러 확인 필요
