+++ src/service/socketIo/constant.ts
@@ -20,3 +20,16 @@ export const TASK = {
   ROOM_UPDATE: "update",
   START_GAME: "start",
 } as const;
+
+export const SOCKET_RESPONSE = {
+  CODE: {
+    RESULT_SUCCESS: 200,
+    /** 해당 방에 없는 사람을 exit 요청하는 경우 */
+    INVALID_REQUEST: 400,
+    PERMISSION_DENIED: 403,
+    NO_RESOURCE: 404,
+    /** 방에 다찾는데 join  */
+    FULL_ROOM_ERROR: 423,
+    INTERNAL_SERVER_ERROR: 500,
+  },
+};
